@@ -4,7 +4,7 @@
           
           <table class="table table_summary">
             <tbody>
-              @foreach(\App\Cart::where('user_id',Auth::id())->orderBy('id')->get() as $n=>$cart_item)
+              @foreach(\App\Models\Cart::where('user_id',Auth::id())->orderBy('id')->get() as $n=>$cart_item)
               <tr>
                 <td><a href="{{URL::to('delete_item/'.$cart_item->id)}}" class="remove_item"><i class="icon_minus_alt"></i></a> <strong>{{$cart_item->quantity}}x</strong> {{$cart_item->item_name}} </td>
                 <td><strong class="pull-right">{{$cart_item->item_price}}</strong></td>
