@@ -36,11 +36,20 @@
                <!-- <div class="avatar">
                     <img src="images/guy.jpg" alt="guy" class="img-circle img-responsive"/>
                 </div>-->
-                {!! Form::open(array('url' => 'admin/login','class'=>'','id'=>'loginform','role'=>'form')) !!}
+                <form action="{{url('admin/login')}}" class="" id="loginform" method="POST">
+                    @csrf
+                    @method('POST')
                     <div class="panel-body">
                     	
                     	<div class="message">
-												<!--{!! Html::ul($errors->all(), array('class'=>'alert alert-danger errors')) !!}-->
+												@if ($errors->any())
+    <ul class="alert alert-danger errors">
+        @foreach ($errors->all() as $error)
+            <li>{{ $error }}</li>
+        @endforeach
+    </ul>
+@endif
+
 							                    	@if (count($errors) > 0)
 											    <div class="alert alert-danger">
 											    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
@@ -56,7 +65,7 @@
 							                    </div>
                     	
                         <div class="form-group">
-                            <label for="email">Email address</label>
+                            <label for="email">Email addresss</label>
                             <input type="email" class="form-control" name="email" id="email" placeholder="Enter email">
                         </div>
                         <div class="form-group margin-none">
@@ -79,7 +88,7 @@
                     <div class="form-group text-center">
                         <button type="submit" class="btn btn-primary">Login <i class="md md-lock-open"></i></button>
                     </div>
-                {!! Form::close() !!} 
+                </form>
             </div>
            
         </div>
