@@ -26,10 +26,12 @@
 				    </div>
 	@endif
    
-   	<div class="panel panel-default">
+   	    <div class="panel panel-default">
             <div class="panel-body">
-                {!! Form::open(array('url' => array('admin/restaurants/view/'.$restaurant_id.'/categories/addcategory'),'class'=>'form-horizontal padding-15','name'=>'category_form','id'=>'category_form','role'=>'form','enctype' => 'multipart/form-data')) !!} 
-                
+                <form action="{{ url('admin/restaurants/view/'.$restaurant_id.'/categories/addcategory') }}" method="POST" class="form-horizontal padding-15" name="category_form" id="category_form" enctype="multipart/form-data">
+                    @csrf
+                    @method('POST')
+
                 <input type="hidden" name="restaurant_id" value="{{$restaurant_id}}">
                 <input type="hidden" name="id" value="{{ isset($cat->id) ? $cat->id : null }}">
                 <div class="form-group">
@@ -46,7 +48,7 @@
                     </div>
                 </div>
                 
-                {!! Form::close() !!} 
+                </form> 
             </div>
         </div>
    
