@@ -36,7 +36,10 @@
                <!-- <div class="avatar">
                     <img src="images/guy.jpg" alt="guy" class="img-circle img-responsive"/>
                 </div>-->
-                {!! Form::open(array('url' => 'admin/password/email','class'=>'','id'=>'passwordform','role'=>'form')) !!}
+            
+                <form action="{{ url('admin/password') }}" method="POST" class="" id="passwordform">
+                    @csrf
+                    @method('POST')
                     <div class="panel-body">
                     	
                     	@if(Session::has('flash_message'))
@@ -47,7 +50,14 @@
 				    </div>
 				@endif
                     	<div class="message">
-												<!--{!! Html::ul($errors->all(), array('class'=>'alert alert-danger errors')) !!}-->
+												@if ($errors->any())
+    <ul class="alert alert-danger errors">
+        @foreach ($errors->all() as $error)
+            <li>{{ $error }}</li>
+        @endforeach
+    </ul>
+@endif
+
 							                    	@if (count($errors) > 0)
 											    <div class="alert alert-danger">
 											    <button type="button" class="close" data-dismiss="alert" aria-label="Close">
@@ -75,7 +85,7 @@
                     <div class="form-group text-center">
                         <button type="submit" class="btn btn-primary"> Send Reset Link <i class="md md-lock-open"></i></button>
                     </div>
-                {!! Form::close() !!} 
+                </form>
             </div>
            
         </div>
